Add mathematica code
